Move numerif function defitions into expr.numeric_expr
